Move terraform.Load() to statemgmt.Load()
Pure move, no functional changes.
