+++ bundle/deploy/terraform/convert.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/databricks/cli/bundle/config"
 	"github.com/databricks/cli/bundle/config/resources"
 	"github.com/databricks/cli/bundle/deploy/terraform/tfdyn"
 	"github.com/databricks/cli/bundle/internal/tf/schema"
@@ -76,66 +75,4 @@ func BundleToTerraformWithDynValue(ctx context.Context, root dyn.Value) (*schema
 	}
 
 	return tfroot, nil
-}
-
-func ensureMap(v dyn.Value, path dyn.Path) (dyn.Value, error) {
-	item, _ := dyn.GetByPath(v, path)
-	if !item.IsValid() {
-		var err error
-		v, err = dyn.SetByPath(v, path, dyn.V(dyn.NewMapping()))
-		if err != nil {
-			return dyn.InvalidValue, fmt.Errorf("internal error: failed to create %s: %s", path, err)
-		}
-	}
-	return v, nil
-}
-
-func TerraformToBundle(ctx context.Context, state ExportedResourcesMap, config *config.Root) error {
-	return config.Mutate(func(v dyn.Value) (dyn.Value, error) {
-		var err error
-		v, err = ensureMap(v, dyn.Path{dyn.Key("resources")})
-		if err != nil {
-			return v, err
-		}
-
-		for groupName, group := range state {
-			var err error
-			v, err = ensureMap(v, dyn.Path{dyn.Key("resources"), dyn.Key(groupName)})
-			if err != nil {
-				return v, err
-			}
-
-			for resourceName, attrs := range group {
-				path := dyn.Path{dyn.Key("resources"), dyn.Key(groupName), dyn.Key(resourceName)}
-				resource, err := dyn.GetByPath(v, path)
-				if !resource.IsValid() {
-					m := dyn.NewMapping()
-					m.SetLoc("id", nil, dyn.V(attrs.ID))
-					m.SetLoc("modified_status", nil, dyn.V(resources.ModifiedStatusDeleted))
-					v, err = dyn.SetByPath(v, path, dyn.V(m))
-					if err != nil {
-						return dyn.InvalidValue, err
-					}
-				} else if err != nil {
-					return dyn.InvalidValue, err
-				} else {
-					v, err = dyn.SetByPath(v, dyn.Path{dyn.Key("resources"), dyn.Key(groupName), dyn.Key(resourceName), dyn.Key("id")}, dyn.V(attrs.ID))
-					if err != nil {
-						return dyn.InvalidValue, err
-					}
-				}
-			}
-		}
-
-		return dyn.MapByPattern(v, dyn.Pattern{dyn.Key("resources"), dyn.AnyKey(), dyn.AnyKey()}, func(p dyn.Path, inner dyn.Value) (dyn.Value, error) {
-			idPath := dyn.Path{dyn.Key("id")}
-			statusPath := dyn.Path{dyn.Key("modified_status")}
-			id, _ := dyn.GetByPath(inner, idPath)
-			status, _ := dyn.GetByPath(inner, statusPath)
-			if !id.IsValid() && !status.IsValid() {
-				return dyn.SetByPath(inner, statusPath, dyn.V(resources.ModifiedStatusCreated))
-			}
-			return inner, nil
-		})
-	})
 }
